refactor: rename stage -> setupStage
