+++ packages/core/src/shared/lsp/lspResolver.ts
@@ -12,7 +12,7 @@ import AdmZip from 'adm-zip'
 import { TargetContent, logger, LspResult, LspVersion, Manifest } from './types'
 import { getApplicationSupportFolder } from '../vscode/env'
 import { createHash } from '../crypto'
-import { lspSetupStage, StageResolver, tryStageResolvers } from './utils/stage'
+import { lspSetupStage, StageResolver, tryStageResolvers } from './utils/setupStage'
 import { HttpResourceFetcher } from '../resourcefetcher/httpResourceFetcher'
 
 export class LanguageServerResolver {
